mido: camera: QCamera2: Fix hang during snapshot after flush
Issue:
If snapshot is triggered after flush in HAL 3 sometimes
it is leading to a hang. If there is a blob capture
request by the app for which HAL has queued the setting
in jpeg setting queue but before de-queuing these setting
if flush is called HAL wont flush the jpeg setting queue.
After flush if another blob capture request is triggered
it will fail because HAL would try to use the setting for
the previous jpeg job which were not flushed.

Fix:
Flush the jpeg setting queue when data proc is stopped.
Initialize the jpeg setting queue when data proc is started.

Change-Id: I9b533e75fb9d87984fb9c3c7f4ce34231403b792